pyproject.toml
[tool.poetry]
name = "chamados-ti"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.119.1"}
sqlalchemy = "^2.0.44"
python-jose = "^3.5.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
pydantic-settings = "^2.11.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.14.1"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'frontend']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev chamados_ti/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=chamados_ti -vv'
post_test = 'coverage html'

